Improve pg function prefix consistency
