+++ core/src/main/java/io/questdb/griffin/engine/functions/catalogue/PgGetPartKeyDefFunctionFactory.java
@@ -24,10 +24,28 @@
 
 package io.questdb.griffin.engine.functions.catalogue;
 
-public class PrefixedTxIDCurrentFunctionFactory extends TxIDCurrentFunctionFactory {
+import io.questdb.cairo.CairoConfiguration;
+import io.questdb.cairo.sql.Function;
+import io.questdb.griffin.FunctionFactory;
+import io.questdb.griffin.SqlExecutionContext;
+import io.questdb.griffin.engine.functions.constants.StrConstant;
+import io.questdb.std.IntList;
+import io.questdb.std.ObjList;
+
+public class PgGetPartKeyDefFunctionFactory implements FunctionFactory {
 
     @Override
     public String getSignature() {
-        return "pg_catalog.txid_current()";
+        return "pg_get_partkeydef(I)";
+    }
+
+    @Override
+    public boolean isRuntimeConstant() {
+        return true;
+    }
+
+    @Override
+    public Function newInstance(int position, ObjList<Function> args, IntList argPositions, CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
+        return StrConstant.NULL;
     }
 }
