+++ core/src/main/java/io/questdb/griffin/engine/functions/catalogue/PgGetSITExprFunctionFactory.java
@@ -28,17 +28,15 @@ import io.questdb.cairo.CairoConfiguration;
 import io.questdb.cairo.sql.Function;
 import io.questdb.griffin.FunctionFactory;
 import io.questdb.griffin.SqlExecutionContext;
-import io.questdb.griffin.engine.functions.StrFunction;
 import io.questdb.griffin.engine.functions.constants.StrConstant;
 import io.questdb.std.IntList;
 import io.questdb.std.ObjList;
 
-public class CurrentSchemaBooleanFunctionFactory implements FunctionFactory {
-    private final static StrFunction INSTANCE = new StrConstant("questdb");
+public class PgGetSITExprFunctionFactory implements FunctionFactory {
 
     @Override
     public String getSignature() {
-        return "current_schemas(T)";
+        return "pg_get_expr(SIT)";
     }
 
     @Override
@@ -48,6 +46,6 @@ public class CurrentSchemaBooleanFunctionFactory implements FunctionFactory {
 
     @Override
     public Function newInstance(int position, ObjList<Function> args, IntList argPositions, CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
-        return INSTANCE;
+        return StrConstant.NULL;
     }
 }
