+++ core/src/main/java/io/questdb/griffin/engine/functions/catalogue/PgIsInRecoveryFunctionFactory.java
@@ -24,10 +24,23 @@
 
 package io.questdb.griffin.engine.functions.catalogue;
 
-public class PrefixedTxIDCurrentFunctionFactory extends TxIDCurrentFunctionFactory {
+import io.questdb.cairo.CairoConfiguration;
+import io.questdb.cairo.sql.Function;
+import io.questdb.griffin.FunctionFactory;
+import io.questdb.griffin.SqlExecutionContext;
+import io.questdb.griffin.engine.functions.constants.BooleanConstant;
+import io.questdb.std.IntList;
+import io.questdb.std.ObjList;
+
+public class PgIsInRecoveryFunctionFactory implements FunctionFactory {
 
     @Override
     public String getSignature() {
-        return "pg_catalog.txid_current()";
+        return "pg_is_in_recovery()";
+    }
+
+    @Override
+    public Function newInstance(int position, ObjList<Function> args, IntList argPositions, CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
+        return BooleanConstant.FALSE;
     }
 }
