+++ core/src/main/java/io/questdb/griffin/engine/functions/catalogue/PgLocksFunctionFactory.java
@@ -29,14 +29,14 @@ import io.questdb.cairo.GenericRecordMetadata;
 import io.questdb.cairo.TableColumnMetadata;
 import io.questdb.cairo.sql.RecordMetadata;
 
-public class PgInheritsFunctionFactory extends AbstractEmptyCatalogueFunctionFactory {
-    private static final RecordMetadata METADATA;
+public class PgLocksFunctionFactory extends AbstractEmptyCatalogueFunctionFactory {
+    private final static RecordMetadata METADATA;
 
-    public PgInheritsFunctionFactory() {
-        super("pg_inherits()", METADATA);
+    public PgLocksFunctionFactory() {
+        this("pg_locks()");
     }
 
-    protected PgInheritsFunctionFactory(String signature) {
+    protected PgLocksFunctionFactory(String signature) {
         super(signature, METADATA);
     }
 
@@ -47,9 +47,22 @@ public class PgInheritsFunctionFactory extends AbstractEmptyCatalogueFunctionFac
 
     static {
         final GenericRecordMetadata metadata = new GenericRecordMetadata();
-        metadata.add(new TableColumnMetadata("inhrelid", ColumnType.INT));
-        metadata.add(new TableColumnMetadata("inhparent", ColumnType.INT));
-        metadata.add(new TableColumnMetadata("inhseqno", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("locktype", ColumnType.STRING));
+        metadata.add(new TableColumnMetadata("database", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("relation", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("page", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("tuple", ColumnType.SHORT));
+        metadata.add(new TableColumnMetadata("virtualxid", ColumnType.STRING));
+        metadata.add(new TableColumnMetadata("transactionid", ColumnType.LONG));
+        metadata.add(new TableColumnMetadata("classid", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("objid", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("objsubid", ColumnType.SHORT));
+        metadata.add(new TableColumnMetadata("virtualtransaction", ColumnType.STRING));
+        metadata.add(new TableColumnMetadata("pid", ColumnType.INT));
+        metadata.add(new TableColumnMetadata("mode", ColumnType.STRING));
+        metadata.add(new TableColumnMetadata("granted", ColumnType.BOOLEAN));
+        metadata.add(new TableColumnMetadata("fastpath", ColumnType.BOOLEAN));
+        metadata.add(new TableColumnMetadata("waitstart", ColumnType.TIMESTAMP));
         METADATA = metadata;
     }
 }
