+++ core/src/main/java/module-info.java
@@ -24,7 +24,6 @@
 
 import io.questdb.griffin.FunctionFactory;
 import io.questdb.griffin.engine.functions.cast.CastNullTypeFunctionFactory;
-import io.questdb.griffin.engine.functions.catalogue.PgClassFunctionFactory;
 import io.questdb.griffin.engine.functions.math.*;
 
 open module io.questdb {
@@ -593,15 +592,18 @@ open module io.questdb {
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgAttrDefFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgAttributeFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgAttributeFunctionFactory,
-            PgClassFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgClassFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgClassFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgIndexFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgIndexFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgRolesFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgRolesFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.InformationSchemaFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgTypeFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgDescriptionFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgNamespaceFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgNamespaceFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgLocksFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgLocksFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.IsTableVisibleCatalogueFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.UserByIdCatalogueFunctionFactory,
@@ -609,23 +611,26 @@ open module io.questdb {
             io.questdb.griffin.engine.functions.catalogue.VersionFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.CurrentDatabaseFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentDatabaseFunctionFactory,
-            io.questdb.griffin.engine.functions.catalogue.CurrentSchemaBooleanFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.CurrentSchemasFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentSchemasFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.CurrentSchemaFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentSchemaFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.CurrentUserFunctionFactory,
-            io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentSchemasFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.CursorDereferenceFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgDescriptionFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgInheritsFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgInheritsFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.SessionUserFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgGetPartKeyDefFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetPartKeyDefFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgGetSITExprFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetSITExprFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetSIExprFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgGetSIExprFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.FormatTypeFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgProcFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgRangeFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgGetKeywordsFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetKeywordsFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.TableListFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.WalTableListFunctionFactory,
@@ -633,7 +638,9 @@ open module io.questdb {
             io.questdb.griffin.engine.functions.catalogue.DumpThreadStacksFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.FlushQueryCacheFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedAgeFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.PgIsInRecoveryFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgIsInRecoveryFunctionFactory,
+            io.questdb.griffin.engine.functions.catalogue.TxIDCurrentFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedTxIDCurrentFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PgDatabaseFunctionFactory,
             io.questdb.griffin.engine.functions.catalogue.PrefixedPgDatabaseFunctionFactory,
