+++ core/src/main/resources/META-INF/services/io.questdb.griffin.FunctionFactory
@@ -566,7 +566,9 @@ io.questdb.griffin.engine.functions.catalogue.PgAttributeFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgAttributeFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgClassFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgClassFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgIndexFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgIndexFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgRolesFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgRolesFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.InformationSchemaFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgTypeFunctionFactory
@@ -577,7 +579,9 @@ io.questdb.griffin.engine.functions.catalogue.UserByIdCatalogueFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgTypeFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.VersionFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgNamespaceFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgLocksFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgLocksFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.CurrentSchemasFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentSchemasFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.CursorDereferenceFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgDescriptionFunctionFactory
@@ -587,13 +591,16 @@ io.questdb.griffin.engine.functions.catalogue.SessionUserFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.CurrentSchemaFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedCurrentSchemaFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.CurrentUserFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgGetPartKeyDefFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetPartKeyDefFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetSIExprFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgGetSIExprFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgGetSITExprFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetSITExprFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.FormatTypeFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgProcFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgRangeFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgGetKeywordsFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgGetKeywordsFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.TableListFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.WalTableListFunctionFactory
@@ -601,7 +608,9 @@ io.questdb.griffin.engine.functions.catalogue.DumpMemoryUsageFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.DumpThreadStacksFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.FlushQueryCacheFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedAgeFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.PgIsInRecoveryFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgIsInRecoveryFunctionFactory
+io.questdb.griffin.engine.functions.catalogue.TxIDCurrentFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedTxIDCurrentFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PgDatabaseFunctionFactory
 io.questdb.griffin.engine.functions.catalogue.PrefixedPgDatabaseFunctionFactory
